[MIG] fetchmail_inbox/10.0
