+++ fetchmail_inbox/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 # © 2014-2016 Therp BV <http://therp.nl>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import model
+from . import models
 from . import wizard
