+++ fetchmail_inbox/__manifest__.py
@@ -3,7 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
     "name": "Fetch mail into inbox",
-    "version": "8.0.1.0.0",
+    "version": "10.0.1.0.0",
     "author": "Therp BV",
     "category": "Dependency",
     "depends": [
@@ -14,8 +14,8 @@
         "security/res_groups.xml",
         "wizard/fetchmail_inbox_create_wizard.xml",
         "wizard/fetchmail_inbox_attach_existing_wizard.xml",
-        "view/mail_message.xml",
-        "view/menu.xml",
+        "views/mail_message.xml",
+        "views/menu.xml",
         'security/ir.model.access.csv',
     ],
 }
