+++ fetchmail_inbox/security/res_groups.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record id="group_fetchmail_inbox_manager" model="res.groups">
-            <field name="name">Fetchmail Inbox Manager</field>
-        </record>
-    </data>
-</openerp>
+<odoo>
+    <record id="group_fetchmail_inbox_manager" model="res.groups">
+        <field name="name">Fetchmail Inbox Manager</field>
+        <field name="users" eval="[(4, ref('base.user_root'))]"/>
+    </record>
+</odoo>
