+++ fetchmail_inbox/views/menu.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+    <act_window id="action_inbox"
+                res_model="mail.message"
+                view_mode="tree,form"
+                domain="[('model', '=', 'fetchmail.inbox')]"
+                name="Fetchmail inbox"
+                view_id="tree_mail_message_fetchmail_inbox"
+                />
+    <menuitem id="menu_inbox"
+              action="action_inbox"
+              parent="base.menu_email"
+              groups="group_fetchmail_inbox_manager"
+              />
+</odoo>
