+++ fetchmail_inbox/wizard/fetchmail_inbox_attach_existing_wizard.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 # © 2014-2016 Therp BV <http://therp.nl>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from openerp import _, api, fields, models
-from openerp.exceptions import Warning as UserError
+from odoo import _, api, fields, models
+from odoo.exceptions import UserError
 
 
 class FetchmailInboxAttachExistingWizard(models.TransientModel):
