+++ fetchmail_inbox/wizard/fetchmail_inbox_attach_existing_wizard.xml
@@ -1,23 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record id="form_fetchmail_inbox_attach_existing_wizard" model="ir.ui.view">
-            <field name="model">fetchmail.inbox.attach.existing.wizard</field>
-            <field name="arch" type="xml">
-                <form>
-                    <group>
-                        <field name="mail_id" invisible="True" />
-                        <field name="res_model" invisible="True" />
-                        <field name="res_id" attrs="{'invisible': [('res_model', '=', False)], 'required': [('res_model', '!=', False)]}" />
-                        <field name="res_reference" attrs="{'invisible': [('res_model', '!=', False)], 'required': [('res_model', '=', False)]}" />
-                    </group>
-                    <footer>
-                        <button type="object" string="Attach to this object" name="button_attach" class="oe_highlight" />
-                        or
-                        <button special="cancel" string="Cancel" class="oe_link" />
-                    </footer>
-                </form>
-            </field>
-        </record>
-    </data>
-</openerp>
+<odoo>
+    <record id="form_fetchmail_inbox_attach_existing_wizard" model="ir.ui.view">
+        <field name="model">fetchmail.inbox.attach.existing.wizard</field>
+        <field name="arch" type="xml">
+            <form>
+                <group>
+                    <field name="mail_id" invisible="True" />
+                    <field name="res_model" invisible="True" />
+                    <field name="res_id" attrs="{'invisible': [('res_model', '=', False)], 'required': [('res_model', '!=', False)]}" />
+                    <field name="res_reference" attrs="{'invisible': [('res_model', '!=', False)], 'required': [('res_model', '=', False)]}" />
+                </group>
+                <footer>
+                    <button type="object" string="Attach to this object" name="button_attach" class="oe_highlight" />
+                    or
+                    <button special="cancel" string="Cancel" class="oe_link" />
+                </footer>
+            </form>
+        </field>
+    </record>
+</odoo>
