+++ fetchmail_inbox/wizard/fetchmail_inbox_create_wizard.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # © 2014-2016 Therp BV <http://therp.nl>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from openerp import api, fields, models
+from odoo import api, fields, models
 
 
 class FetchmailInboxCreateWizard(models.TransientModel):
